# CONNECT-HUB-Api-Gateway/cmd/docs/swagger.yaml
definitions:
  models.AdminLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  models.CreatePolicyReq:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
...
  models.UpdatePolicyReq:
    properties:
      content:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: ConnectHub is a professional networking platform.
  title: Go + Gin professional networking platform API Connect Hub
  version: 1.0.0
paths:
  /admin/jobseeker:
    get:
      consumes:
      - application/json
      description: Retrieve the details of a jobseeker based on the provided ID
      parameters:
      - description: Jobseeker ID to get details
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Jobseeker details retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid jobseeker ID or failed to retrieve jobseeker details
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get jobseeker details
      tags:
      - Admin Jobseeker Management
  /admin/jobseeker/all:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of jobseekers
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved jobseekers
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid page number or constraints not satisfied
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve jobseekers due to internal error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get jobseekers
      tags:
      - Admin Jobseeker Management
  /admin/jobseeker/block:
    patch:
      consumes:
      - application/json
      description: Block a jobseeker based on the provided ID
      parameters:
      - description: Jobseeker ID to block
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Jobseeker blocked successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid jobseeker ID or failed to block jobseeker
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Block a jobseeker
      tags:
      - Admin Jobseeker Management
  /admin/jobseeker/unblock:
    patch:
      consumes:
      - application/json
      description: Unblock a jobseeker based on the provided ID
      parameters:
      - description: Jobseeker ID to unblock
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Jobseeker unblocked successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid jobseeker ID or failed to unblock jobseeker
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Unblock a jobseeker
      tags:
      - Admin Jobseeker Management
  /admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate an admin and get access token
      parameters:
      - description: Admin credentials for login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AdminLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Admin login successful
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request or constraints not satisfied
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: cannot authenticate user'
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin login
      tags:
      - Admin Authentication Management
  /admin/policies:
    get:
      consumes:
      - application/json
      description: Retrieves all existing policies
      produces:
      - application/json
      responses:
        "200":
          description: Policies retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to retrieve policies
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Retrieve all policies
      tags:
      - Admin Policy Management
  /admin/policy:
    delete:
      consumes:
      - application/json
      description: Delete a policy based on the provided ID
      parameters:
      - description: Policy ID to delete
        in: query
        name: policy_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Policy deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid policy ID or failed to delete policy
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Delete a policy
      tags:
      - Admin Policy Management
    get:
      consumes:
      - application/json
      description: Retrieves a policy based on the provided ID
      parameters:
      - description: Policy ID to retrieve
        in: query
        name: policy_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Policy retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid policy ID or failed to retrieve policy
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Retrieve a policy
      tags:
      - Admin Policy Management
    post:
      consumes:
      - application/json
      description: Create a new policy based on the provided data
      parameters:
      - description: Policy data
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/models.CreatePolicyReq'
      produces:
      - application/json
      responses:
        "200":
          description: Policy created successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid policy data or failed to create policy
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Create a new policy
      tags:
      - Admin Policy Management
    put:
      consumes:
      - application/json
      description: Update an existing policy based on the provided data
      parameters:
      - description: Updated policy data
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePolicyReq'
      produces:
      - application/json
      responses:
        "200":
          description: Policy updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid policy data or failed to update policy
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Update an existing policy
      tags:
      - Admin Policy Management
  /admin/recruiter:
    get:
      consumes:
      - application/json
      description: Retrieve the details of a recruiter based on the provided ID
      parameters:
      - description: Recruiter ID to get details
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recruiter details retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid recruiter ID or failed to retrieve recruiter details
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get recruiter details
      tags:
      - Admin Recruiter Management
  /admin/recruiter/all:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of recruiters
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved recruiters
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid page number or constraints not satisfied
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve recruiters due to internal error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get recruiters
      tags:
      - Admin Recruiter Management
  /admin/recruiter/block:
    patch:
      consumes:
      - application/json
      description: Block a recruiter based on the provided ID
      parameters:
      - description: Recruiter ID to block
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recruiter blocked successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid recruiter ID or failed to block recruiter
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Block a recruiter
      tags:
      - Admin Recruiter Management
  /admin/recruiter/unblock:
    patch:
      consumes:
      - application/json
      description: Unblock a recruiter based on the provided ID
      parameters:
      - description: Recruiter ID to unblock
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recruiter unblocked successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid recruiter ID or failed to unblock recruiter
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Unblock a recruiter
      tags:
      - Admin Recruiter Management
...
securityDefinitions:
  BearerTokenAuth:
    description: 'For accessing the API, you need to include "Jobseeker","Admin" or
      "Recruiter" before your token in the Authorization header. Example: "Authorization:
      Admin <your_token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"